refactor: move resolve_note/follow_link away from client
