+++ lua/obsidian/client.lua
@@ -17,7 +17,6 @@ local util = require "obsidian.util"
 local search = require "obsidian.search"
 local AsyncExecutor = require("obsidian.async").AsyncExecutor
 local block_on = require("obsidian.async").block_on
-local api = require "obsidian.api"
 local iter = vim.iter
 
 ---@class obsidian.SearchOpts
@@ -294,49 +293,6 @@ Client.find_notes_async = function(self, term, callback, opts)
   end, function(_) end)
 end
 
---- Find non-markdown files in the vault.
----
----@param term string The search term.
----@param opts { search: obsidian.SearchOpts, timeout: integer|? }|?
----
----@return obsidian.Path[]
-Client.find_files = function(self, term, opts)
-  opts = opts or {}
-  return block_on(function(cb)
-    return self:find_files_async(term, cb, opts)
-  end, opts.timeout)
-end
-
---- An async version of `find_files`.
----
----@param term string The search term.
----@param callback fun(paths: obsidian.Path[])
----@param opts { search: obsidian.SearchOpts }|?
-Client.find_files_async = function(self, term, callback, opts)
-  opts = opts or {}
-
-  local matches = {}
-  local tx, rx = channel.oneshot()
-  local on_find_match = function(path_match)
-    matches[#matches + 1] = Path.new(path_match)
-  end
-
-  local on_exit = function(_)
-    tx()
-  end
-
-  local find_opts = self:_prepare_search_opts(opts.search)
-  search.SearchOpts.add_exclude(find_opts, "*.md")
-  find_opts.include_non_markdown = true
-
-  search.find_async(Obsidian.dir, term, find_opts, on_find_match, on_exit)
-
-  async.run(function()
-    rx()
-    return matches
-  end, callback)
-end
-
 --- Resolve the query to a single note if possible, otherwise all close matches are returned.
 --- The 'query' can be a path, filename, note ID, alias, title, etc.
 ---
@@ -347,352 +303,8 @@ end
 Client.resolve_note = function(self, query, opts)
   opts = opts or {}
   return block_on(function(cb)
-    return self:resolve_note_async(query, cb, { notes = opts.notes })
+    return search.resolve_note_async(query, cb, { notes = opts.notes })
   end, opts.timeout)
-end
-
---- An async version of `resolve_note()`.
----
----@param query string
----@param callback fun(...: obsidian.Note)
----@param opts { notes: obsidian.note.LoadOpts|? }|?
----
----@return obsidian.Note|?
-Client.resolve_note_async = function(self, query, callback, opts)
-  opts = opts or {}
-  opts.notes = opts.notes or {}
-  if not opts.notes.max_lines then
-    opts.notes.max_lines = Obsidian.opts.search_max_lines
-  end
-
-  -- Autocompletion for command args will have this format.
-  local note_path, count = string.gsub(query, "^.*  ", "")
-  if count > 0 then
-    ---@type obsidian.Path
-    ---@diagnostic disable-next-line: assign-type-mismatch
-    local full_path = Obsidian.dir / note_path
-    return async.run(function()
-      return Note.from_file_async(full_path, opts.notes)
-    end, callback)
-  end
-
-  -- Query might be a path.
-  local fname = query
-  if not vim.endswith(fname, ".md") then
-    fname = fname .. ".md"
-  end
-
-  local paths_to_check = { Path.new(fname), Obsidian.dir / fname }
-
-  if Obsidian.opts.notes_subdir ~= nil then
-    paths_to_check[#paths_to_check + 1] = Obsidian.dir / Obsidian.opts.notes_subdir / fname
-  end
-
-  if Obsidian.opts.daily_notes.folder ~= nil then
-    paths_to_check[#paths_to_check + 1] = Obsidian.dir / Obsidian.opts.daily_notes.folder / fname
-  end
-
-  if Obsidian.buf_dir ~= nil then
-    paths_to_check[#paths_to_check + 1] = Obsidian.buf_dir / fname
-  end
-
-  for _, path in pairs(paths_to_check) do
-    if path:is_file() then
-      return async.run(function()
-        return Note.from_file_async(path, opts.notes)
-      end, callback)
-    end
-  end
-
-  self:find_notes_async(query, function(results)
-    local query_lwr = string.lower(query)
-
-    -- We'll gather both exact matches (of ID, filename, and aliases) and fuzzy matches.
-    -- If we end up with any exact matches, we'll return those. Otherwise we fall back to fuzzy
-    -- matches.
-    ---@type obsidian.Note[]
-    local exact_matches = {}
-    ---@type obsidian.Note[]
-    local fuzzy_matches = {}
-
-    for note in iter(results) do
-      ---@cast note obsidian.Note
-
-      local reference_ids = note:reference_ids { lowercase = true }
-
-      -- Check for exact match.
-      if vim.list_contains(reference_ids, query_lwr) then
-        table.insert(exact_matches, note)
-      else
-        -- Fall back to fuzzy match.
-        for ref_id in iter(reference_ids) do
-          if util.string_contains(ref_id, query_lwr) then
-            table.insert(fuzzy_matches, note)
-            break
-          end
-        end
-      end
-    end
-
-    if #exact_matches > 0 then
-      return callback(unpack(exact_matches))
-    else
-      return callback(unpack(fuzzy_matches))
-    end
-  end, { search = { sort = true, ignore_case = true }, notes = opts.notes })
-end
-
---- Same as `resolve_note_async` but opens a picker to choose a single note when
---- there are multiple matches.
----
----@param query string
----@param callback fun(obsidian.Note)
----@param opts { notes: obsidian.note.LoadOpts|?, prompt_title: string|? }|?
----
----@return obsidian.Note|?
-Client.resolve_note_async_with_picker_fallback = function(self, query, callback, opts)
-  opts = opts or {}
-
-  self:resolve_note_async(query, function(...)
-    local notes = { ... }
-
-    if #notes == 0 then
-      log.err("No notes matching '%s'", query)
-      return
-    elseif #notes == 1 then
-      return callback(notes[1])
-    end
-
-    -- Fall back to picker.
-    vim.schedule(function()
-      -- Otherwise run the preferred picker to search for notes.
-      local picker = Obsidian.picker
-      if not picker then
-        log.err("Found multiple notes matching '%s', but no picker is configured", query)
-        return
-      end
-
-      picker:pick_note(notes, {
-        prompt_title = opts.prompt_title,
-        callback = callback,
-      })
-    end)
-  end, { notes = opts.notes })
-end
-
----@class obsidian.ResolveLinkResult
----
----@field location string
----@field name string
----@field link_type obsidian.search.RefTypes
----@field path obsidian.Path|?
----@field note obsidian.Note|?
----@field url string|?
----@field line integer|?
----@field col integer|?
----@field anchor obsidian.note.HeaderAnchor|?
----@field block obsidian.note.Block|?
-
---- Resolve a link. If the link argument is `nil` we attempt to resolve a link under the cursor.
----
----@param link string|?
----@param callback fun(...: obsidian.ResolveLinkResult)
-Client.resolve_link_async = function(self, link, callback)
-  local location, name, link_type
-  if link then
-    location, name, link_type = util.parse_link(link, { include_naked_urls = true, include_file_urls = true })
-  else
-    location, name, link_type = api.parse_cursor_link { include_naked_urls = true, include_file_urls = true }
-  end
-
-  if location == nil or name == nil or link_type == nil then
-    return callback()
-  end
-
-  ---@type obsidian.ResolveLinkResult
-  local res = { location = location, name = name, link_type = link_type }
-
-  if util.is_url(location) then
-    res.url = location
-    return callback(res)
-  end
-
-  -- The Obsidian app will follow URL-encoded links, so we should to.
-  location = vim.uri_decode(location)
-
-  -- Remove block links from the end if there are any.
-  -- TODO: handle block links.
-  ---@type string|?
-  local block_link
-  location, block_link = util.strip_block_links(location)
-
-  -- Remove anchor links from the end if there are any.
-  ---@type string|?
-  local anchor_link
-  location, anchor_link = util.strip_anchor_links(location)
-
-  --- Finalize the `obsidian.ResolveLinkResult` for a note while resolving block or anchor link to line.
-  ---
-  ---@param note obsidian.Note
-  ---@return obsidian.ResolveLinkResult
-  local function finalize_result(note)
-    ---@type integer|?, obsidian.note.Block|?, obsidian.note.HeaderAnchor|?
-    local line, block_match, anchor_match
-    if block_link ~= nil then
-      block_match = note:resolve_block(block_link)
-      if block_match then
-        line = block_match.line
-      end
-    elseif anchor_link ~= nil then
-      anchor_match = note:resolve_anchor_link(anchor_link)
-      if anchor_match then
-        line = anchor_match.line
-      end
-    end
-
-    return vim.tbl_extend(
-      "force",
-      res,
-      { path = note.path, note = note, line = line, block = block_match, anchor = anchor_match }
-    )
-  end
-
-  ---@type obsidian.note.LoadOpts
-  local load_opts = {
-    collect_anchor_links = anchor_link and true or false,
-    collect_blocks = block_link and true or false,
-    max_lines = Obsidian.opts.search_max_lines,
-  }
-
-  -- Assume 'location' is current buffer path if empty, like for TOCs.
-  if string.len(location) == 0 then
-    res.location = vim.api.nvim_buf_get_name(0)
-    local note = Note.from_buffer(0, load_opts)
-    return callback(finalize_result(note))
-  end
-
-  res.location = location
-
-  self:resolve_note_async(location, function(...)
-    local notes = { ... }
-
-    if #notes == 0 then
-      local path = Path.new(location)
-      if path:exists() then
-        res.path = path
-        return callback(res)
-      else
-        return callback(res)
-      end
-    end
-
-    local matches = {}
-    for _, note in ipairs(notes) do
-      table.insert(matches, finalize_result(note))
-    end
-
-    return callback(unpack(matches))
-  end, { notes = load_opts })
-end
-
---- Follow a link. If the link argument is `nil` we attempt to follow a link under the cursor.
----
----@param link string|?
----@param opts { open_strategy: obsidian.config.OpenStrategy|? }|?
-Client.follow_link_async = function(self, link, opts)
-  opts = opts and opts or {}
-
-  self:resolve_link_async(link, function(...)
-    local results = { ... }
-
-    if #results == 0 then
-      return
-    end
-
-    ---@param res obsidian.ResolveLinkResult
-    local function follow_link(res)
-      if res.url ~= nil then
-        Obsidian.opts.follow_url_func(res.url)
-        return
-      end
-
-      if util.is_img(res.location) then
-        local path = Obsidian.dir / res.location
-        Obsidian.opts.follow_img_func(tostring(path))
-        return
-      end
-
-      if res.note ~= nil then
-        -- Go to resolved note.
-        return res.note:open { line = res.line, col = res.col, open_strategy = opts.open_strategy }
-      end
-
-      if res.link_type == search.RefTypes.Wiki or res.link_type == search.RefTypes.WikiWithAlias then
-        -- Prompt to create a new note.
-        if api.confirm("Create new note '" .. res.location .. "'?") then
-          -- Create a new note.
-          ---@type string|?, string[]
-          local id, aliases
-          if res.name == res.location then
-            aliases = {}
-          else
-            aliases = { res.name }
-            id = res.location
-          end
-
-          local note = Note.create { title = res.name, id = id, aliases = aliases }
-          return note:open {
-            open_strategy = opts.open_strategy,
-            callback = function(bufnr)
-              note:write_to_buffer { bufnr = bufnr }
-            end,
-          }
-        else
-          log.warn "Aborted"
-          return
-        end
-      end
-
-      return log.err("Failed to resolve file '" .. res.location .. "'")
-    end
-
-    if #results == 1 then
-      return vim.schedule(function()
-        follow_link(results[1])
-      end)
-    else
-      return vim.schedule(function()
-        local picker = Obsidian.picker
-        if not picker then
-          log.err("Found multiple matches to '%s', but no picker is configured", link)
-          return
-        end
-
-        ---@type obsidian.PickerEntry[]
-        local entries = {}
-        for _, res in ipairs(results) do
-          local icon, icon_hl
-          if res.url ~= nil then
-            icon, icon_hl = api.get_icon(res.url)
-          end
-          table.insert(entries, {
-            value = res,
-            display = res.name,
-            filename = res.path and tostring(res.path) or nil,
-            icon = icon,
-            icon_hl = icon_hl,
-          })
-        end
-
-        picker:pick(entries, {
-          prompt_title = "Follow link",
-          callback = function(res)
-            follow_link(res)
-          end,
-        })
-      end)
-    end
-  end)
 end
 
 ---@class obsidian.TagLocation
