+++ lua/obsidian/commands/follow_link.lua
@@ -1,10 +1,11 @@
----@param client obsidian.Client
+local api = require "obsidian.api"
+
 ---@param data CommandArgs
-return function(client, data)
+return function(_, data)
   local opts = {}
   if data.args and string.len(data.args) > 0 then
     opts.open_strategy = data.args
   end
 
-  client:follow_link_async(nil, opts)
+  api.follow_link(nil, opts)
 end
